rename to ProjectUpdatesBlock
