+++ src/components/views/homepage/HomeIndex.tsx
@@ -8,7 +8,8 @@ import WhyGiveth from '@/components/views/homepage/whyGiveth';
 import CampaignBlock from '@/components/views/homepage/CampaignBlock';
 import IntroBlock from './introBlock';
 import VideoBlock from './videoBlock';
-import { ProjectActivity } from './projectAtivityBlock/ProjectActivity';
+import { ProjectActivity } from './projectUpdatesBlock/ProjectActivity';
+import { ProjectUpdatesBlock } from './projectUpdatesBlock/ProjectUpdatesBlock';
 
 interface IHomeView {
 	projects: IProject[];
@@ -25,6 +26,7 @@ const HomeIndex: FC<IHomeView> = ({ projects }) => {
 			<WhyGiveth />
 			<Separator />
 			<VideoBlock />
+			<ProjectUpdatesBlock projects={projects} />
 			<ProjectActivity />
 			<Separator />
 			<HomeFromBlog />
