+++ src/components/views/homepage/projectUpdatesBlock/ProjectActivitySlide.tsx
